--refactor HomeScreen
